third_party/blink/web_tests/fast/html/imports/import-dynamic-expected.txt

CONSOLE WARNING: line 16: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This test ensures that script-added import doesn't block scripts

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

In onload handler - This should be called later.
PASS importIsLoaded is true
FAIL scriptIsExecuted should be true. Was false.
In script tag - This should be called first.
PASS importIsLoaded is true
PASS successfullyParsed is true

TEST COMPLETE
